Added tasks gantt chart calendar scope slider
